requirements/common.txt
Django>=5.1.4,<5.2
django-cors-headers>=4.6.0,<4.7
django-environ>=0.11.2,<0.12
djangorestframework>=3.15.2,<3.16
djangorestframework-simplejwt>=5.3.1,<5.4
drf-spectacular>=0.28.0,<0.29
psycopg2>=2.9.10,<2.10
